refactor: create /hooks/challenge directory
