+++ src/pages/challenges/[id]/teams/[id]/index.tsx
@@ -7,7 +7,7 @@ import MemberCount from '@/components/common/teams/MemberCount'
 import Overview from '@/components/common/teams/Overview'
 import PropertyList from '@/components/common/teams/PropertyList'
 import { Button } from '@/components/ui/button'
-import useChallengesTeam from '@/hooks/useChallengesTeam'
+import useChallengesTeam from '@/hooks/challenge/useChallengesTeam'
 import { toast } from 'sonner'
 import { Dialog, DialogContent, DialogDescription } from '@/components/ui/dialog'
 import { useMutation, useQueryClient } from '@tanstack/react-query'
