+++ src/pages/challenges/[id]/teams/index.tsx
@@ -4,7 +4,7 @@ import PageContainer from '@/components/common/PageContainer'
 import PageHeaderSection from '@/components/common/PageHeaderSection'
 import PageTitle from '@/components/common/PageTitle'
 import { Button } from '@/components/ui/button'
-import useChallenge from '@/hooks/useChallenge'
+import useChallenge from '@/hooks/challenge/useChallenge'
 import { useQuery } from '@tanstack/react-query'
 import { useNavigate, useParams } from 'react-router-dom'
 import { SquareCheckBig } from 'lucide-react'
